extract partitionBlocksByType into dedicated util function and update tests
