+++ apps/src/blockly/addons/cdoXml.js
@@ -1,4 +1,5 @@
 import {PROCEDURE_DEFINITION_TYPES} from '../constants';
+import {partitionBlocksByType} from './cdoUtils';
 
 export default function initializeBlocklyXml(blocklyWrapper) {
   // Clear xml namespace
@@ -126,26 +127,4 @@ export function getPartitionedBlockElements(xml, prioritizedBlockTypes) {
     prioritizedBlockTypes
   );
   return partitionedBlockElements;
-}
-
-/**
- * Partitions blocks of the specified types to the front of the list.
- *
- * @param {Element[]} blockElements - An array of block elements to be partitioned.
- * @param {string[]} prioritizedBlockTypes - An array of strings representing block types.
- *    These types are moved to the front of the list while otherwise maintaining order.
- * @returns {Element[]} A new array of block elements partitioned based on their types.
- */
-export function partitionBlocksByType(blockElements, prioritizedBlockTypes) {
-  const prioritizedBlocks = [];
-  const remainingBlocks = [];
-
-  blockElements.forEach(block => {
-    const blockType = block.getAttribute('type');
-    prioritizedBlockTypes.includes(blockType)
-      ? prioritizedBlocks.push(block)
-      : remainingBlocks.push(block);
-  });
-
-  return [...prioritizedBlocks, ...remainingBlocks];
 }
